Maximum coherence and partial coherence matrix

Computes the coherence or partial coherence matrix for a multivatiate (multitype) point pattern.

Usage

coherence(sp.est, ppp, type = "partial", all = FALSE)

Arguments

sp.est

List. The kernel spectral estimate from periodogram_smooth().

ppp

A point pattern of class "ppp".

type

If "partial" (default), calculate the partial coherence. If "normal", calculate the coherence.

all

Logical. If TRUE, extract the maximum (partial) coherence across all frequencies, which is not recommended.

Value

A matrix with each entry storing the maximum (partial) coherence of the two individual point processes.

Examples

library(spatstat)
#> Loading required package: spatstat.data
#> Loading required package: spatstat.univar
#> spatstat.univar 3.1-1
#> Loading required package: spatstat.geom
#> spatstat.geom 3.3-5
#> Loading required package: spatstat.random
#> spatstat.random 3.3-2
#> Loading required package: spatstat.explore
#> Loading required package: nlme
#> spatstat.explore 3.3-4
#> Loading required package: spatstat.model
#> Loading required package: rpart
#> spatstat.model 3.3-4
#> Loading required package: spatstat.linnet
#> spatstat.linnet 3.2-5
#> 
#> spatstat 3.3-1 
#> For an introduction to spatstat, type ‘beginner’ 
lam <- function(x, y, m) {(x^2 + y) * ifelse(m == "A", 2, 1)}
spp <- rmpoispp(lambda = lam, win = square(5), types = c("A","B"))

# Compute kernel spectral estimator with fitted intensity by log-linear model:
# with Cartesian coordinates
spectra <- periodogram_smooth(spp, inten.formula = "~ x + y", bandwidth = 1.2)
coh.partial <- coherence(spectra, spp) # Compute the maximum partial coherence
#> Number of frequencies to pick the maximum partial coherence: 9 (this value should not be too small).
attr(coh.partial, "CohTable") # Print the partial coherence values for all frequencies
#>             omega1    omega2             
#> result.1 -2.827433 -2.827433 9.581547e-05
#> result.2  0.000000 -2.827433 8.064383e-04
#> result.3  2.827433 -2.827433 3.263729e-04
#> result.4 -2.827433  0.000000 1.218888e-04
#> result.5  0.000000  0.000000 2.185532e-05
#> result.6  2.827433  0.000000 1.218888e-04
#> result.7 -2.827433  2.827433 3.263729e-04
#> result.8  0.000000  2.827433 8.064383e-04
#> result.9  2.827433  2.827433 9.581547e-05