Frequency grid generator

Frequency grid generator

Usage

generate_freq(A1, A2, ext.factor = NULL, return.comb = FALSE, endpt = 1.5)

Arguments

A1, A2

Side lengths of the observational window.

ext.factor

A positive value indicating the extension factor of frequency. If NULL, no extension is conducted.

return.comb

Logical. If TURE, also return the data.frame.

endpt

A positive value indicating the scale factor of the endpoint frequency.

Value

A list of two frequency vectors. If return.comb = TRUE, also return a data.frame containing all combinations of frequencies.

Details

This function generates the frequency grid which is used in periodogram() and periodogram_smooth(). In periodogram_smooth(), to correct the edge effect, the ext.factor argument in generate_freq() is set to 2 by default. Suppose the original frequency ranges from -1.5*π to 1.5*π. Then the extended frequency ranges from -1.5*π*ext.factor to 1.5*π*ext.factor.

Examples

generate_freq(A1 = 8, A2 = 5, return.comb = TRUE)
#> $omega1
#>  [1] -4.7123890 -4.1233404 -3.5342917 -2.9452431 -2.3561945 -1.7671459
#>  [7] -1.1780972 -0.5890486  0.0000000  0.5890486  1.1780972  1.7671459
#> [13]  2.3561945  2.9452431  3.5342917  4.1233404  4.7123890
#> 
#> $omega2
#>  [1] -4.7123890 -3.7699112 -2.8274334 -1.8849556 -0.9424778  0.0000000
#>  [7]  0.9424778  1.8849556  2.8274334  3.7699112  4.7123890
#> 
#> $omega.comb
#>            omega1     omega2 Density
#>   [1,] -4.7123890 -4.7123890      NA
#>   [2,] -4.1233404 -4.7123890      NA
#>   [3,] -3.5342917 -4.7123890      NA
#>   [4,] -2.9452431 -4.7123890      NA
#>   [5,] -2.3561945 -4.7123890      NA
#>   [6,] -1.7671459 -4.7123890      NA
#>   [7,] -1.1780972 -4.7123890      NA
#>   [8,] -0.5890486 -4.7123890      NA
#>   [9,]  0.0000000 -4.7123890      NA
#>  [10,]  0.5890486 -4.7123890      NA
#>  [11,]  1.1780972 -4.7123890      NA
#>  [12,]  1.7671459 -4.7123890      NA
#>  [13,]  2.3561945 -4.7123890      NA
#>  [14,]  2.9452431 -4.7123890      NA
#>  [15,]  3.5342917 -4.7123890      NA
#>  [16,]  4.1233404 -4.7123890      NA
#>  [17,]  4.7123890 -4.7123890      NA
#>  [18,] -4.7123890 -3.7699112      NA
#>  [19,] -4.1233404 -3.7699112      NA
#>  [20,] -3.5342917 -3.7699112      NA
#>  [21,] -2.9452431 -3.7699112      NA
#>  [22,] -2.3561945 -3.7699112      NA
#>  [23,] -1.7671459 -3.7699112      NA
#>  [24,] -1.1780972 -3.7699112      NA
#>  [25,] -0.5890486 -3.7699112      NA
#>  [26,]  0.0000000 -3.7699112      NA
#>  [27,]  0.5890486 -3.7699112      NA
#>  [28,]  1.1780972 -3.7699112      NA
#>  [29,]  1.7671459 -3.7699112      NA
#>  [30,]  2.3561945 -3.7699112      NA
#>  [31,]  2.9452431 -3.7699112      NA
#>  [32,]  3.5342917 -3.7699112      NA
#>  [33,]  4.1233404 -3.7699112      NA
#>  [34,]  4.7123890 -3.7699112      NA
#>  [35,] -4.7123890 -2.8274334      NA
#>  [36,] -4.1233404 -2.8274334      NA
#>  [37,] -3.5342917 -2.8274334      NA
#>  [38,] -2.9452431 -2.8274334      NA
#>  [39,] -2.3561945 -2.8274334      NA
#>  [40,] -1.7671459 -2.8274334      NA
#>  [41,] -1.1780972 -2.8274334      NA
#>  [42,] -0.5890486 -2.8274334      NA
#>  [43,]  0.0000000 -2.8274334      NA
#>  [44,]  0.5890486 -2.8274334      NA
#>  [45,]  1.1780972 -2.8274334      NA
#>  [46,]  1.7671459 -2.8274334      NA
#>  [47,]  2.3561945 -2.8274334      NA
#>  [48,]  2.9452431 -2.8274334      NA
#>  [49,]  3.5342917 -2.8274334      NA
#>  [50,]  4.1233404 -2.8274334      NA
#>  [51,]  4.7123890 -2.8274334      NA
#>  [52,] -4.7123890 -1.8849556      NA
#>  [53,] -4.1233404 -1.8849556      NA
#>  [54,] -3.5342917 -1.8849556      NA
#>  [55,] -2.9452431 -1.8849556      NA
#>  [56,] -2.3561945 -1.8849556      NA
#>  [57,] -1.7671459 -1.8849556      NA
#>  [58,] -1.1780972 -1.8849556      NA
#>  [59,] -0.5890486 -1.8849556      NA
#>  [60,]  0.0000000 -1.8849556      NA
#>  [61,]  0.5890486 -1.8849556      NA
#>  [62,]  1.1780972 -1.8849556      NA
#>  [63,]  1.7671459 -1.8849556      NA
#>  [64,]  2.3561945 -1.8849556      NA
#>  [65,]  2.9452431 -1.8849556      NA
#>  [66,]  3.5342917 -1.8849556      NA
#>  [67,]  4.1233404 -1.8849556      NA
#>  [68,]  4.7123890 -1.8849556      NA
#>  [69,] -4.7123890 -0.9424778      NA
#>  [70,] -4.1233404 -0.9424778      NA
#>  [71,] -3.5342917 -0.9424778      NA
#>  [72,] -2.9452431 -0.9424778      NA
#>  [73,] -2.3561945 -0.9424778      NA
#>  [74,] -1.7671459 -0.9424778      NA
#>  [75,] -1.1780972 -0.9424778      NA
#>  [76,] -0.5890486 -0.9424778      NA
#>  [77,]  0.0000000 -0.9424778      NA
#>  [78,]  0.5890486 -0.9424778      NA
#>  [79,]  1.1780972 -0.9424778      NA
#>  [80,]  1.7671459 -0.9424778      NA
#>  [81,]  2.3561945 -0.9424778      NA
#>  [82,]  2.9452431 -0.9424778      NA
#>  [83,]  3.5342917 -0.9424778      NA
#>  [84,]  4.1233404 -0.9424778      NA
#>  [85,]  4.7123890 -0.9424778      NA
#>  [86,] -4.7123890  0.0000000      NA
#>  [87,] -4.1233404  0.0000000      NA
#>  [88,] -3.5342917  0.0000000      NA
#>  [89,] -2.9452431  0.0000000      NA
#>  [90,] -2.3561945  0.0000000      NA
#>  [91,] -1.7671459  0.0000000      NA
#>  [92,] -1.1780972  0.0000000      NA
#>  [93,] -0.5890486  0.0000000      NA
#>  [94,]  0.0000000  0.0000000      NA
#>  [95,]  0.5890486  0.0000000      NA
#>  [96,]  1.1780972  0.0000000      NA
#>  [97,]  1.7671459  0.0000000      NA
#>  [98,]  2.3561945  0.0000000      NA
#>  [99,]  2.9452431  0.0000000      NA
#> [100,]  3.5342917  0.0000000      NA
#> [101,]  4.1233404  0.0000000      NA
#> [102,]  4.7123890  0.0000000      NA
#> [103,] -4.7123890  0.9424778      NA
#> [104,] -4.1233404  0.9424778      NA
#> [105,] -3.5342917  0.9424778      NA
#> [106,] -2.9452431  0.9424778      NA
#> [107,] -2.3561945  0.9424778      NA
#> [108,] -1.7671459  0.9424778      NA
#> [109,] -1.1780972  0.9424778      NA
#> [110,] -0.5890486  0.9424778      NA
#> [111,]  0.0000000  0.9424778      NA
#> [112,]  0.5890486  0.9424778      NA
#> [113,]  1.1780972  0.9424778      NA
#> [114,]  1.7671459  0.9424778      NA
#> [115,]  2.3561945  0.9424778      NA
#> [116,]  2.9452431  0.9424778      NA
#> [117,]  3.5342917  0.9424778      NA
#> [118,]  4.1233404  0.9424778      NA
#> [119,]  4.7123890  0.9424778      NA
#> [120,] -4.7123890  1.8849556      NA
#> [121,] -4.1233404  1.8849556      NA
#> [122,] -3.5342917  1.8849556      NA
#> [123,] -2.9452431  1.8849556      NA
#> [124,] -2.3561945  1.8849556      NA
#> [125,] -1.7671459  1.8849556      NA
#> [126,] -1.1780972  1.8849556      NA
#> [127,] -0.5890486  1.8849556      NA
#> [128,]  0.0000000  1.8849556      NA
#> [129,]  0.5890486  1.8849556      NA
#> [130,]  1.1780972  1.8849556      NA
#> [131,]  1.7671459  1.8849556      NA
#> [132,]  2.3561945  1.8849556      NA
#> [133,]  2.9452431  1.8849556      NA
#> [134,]  3.5342917  1.8849556      NA
#> [135,]  4.1233404  1.8849556      NA
#> [136,]  4.7123890  1.8849556      NA
#> [137,] -4.7123890  2.8274334      NA
#> [138,] -4.1233404  2.8274334      NA
#> [139,] -3.5342917  2.8274334      NA
#> [140,] -2.9452431  2.8274334      NA
#> [141,] -2.3561945  2.8274334      NA
#> [142,] -1.7671459  2.8274334      NA
#> [143,] -1.1780972  2.8274334      NA
#> [144,] -0.5890486  2.8274334      NA
#> [145,]  0.0000000  2.8274334      NA
#> [146,]  0.5890486  2.8274334      NA
#> [147,]  1.1780972  2.8274334      NA
#> [148,]  1.7671459  2.8274334      NA
#> [149,]  2.3561945  2.8274334      NA
#> [150,]  2.9452431  2.8274334      NA
#> [151,]  3.5342917  2.8274334      NA
#> [152,]  4.1233404  2.8274334      NA
#> [153,]  4.7123890  2.8274334      NA
#> [154,] -4.7123890  3.7699112      NA
#> [155,] -4.1233404  3.7699112      NA
#> [156,] -3.5342917  3.7699112      NA
#> [157,] -2.9452431  3.7699112      NA
#> [158,] -2.3561945  3.7699112      NA
#> [159,] -1.7671459  3.7699112      NA
#> [160,] -1.1780972  3.7699112      NA
#> [161,] -0.5890486  3.7699112      NA
#> [162,]  0.0000000  3.7699112      NA
#> [163,]  0.5890486  3.7699112      NA
#> [164,]  1.1780972  3.7699112      NA
#> [165,]  1.7671459  3.7699112      NA
#> [166,]  2.3561945  3.7699112      NA
#> [167,]  2.9452431  3.7699112      NA
#> [168,]  3.5342917  3.7699112      NA
#> [169,]  4.1233404  3.7699112      NA
#> [170,]  4.7123890  3.7699112      NA
#> [171,] -4.7123890  4.7123890      NA
#> [172,] -4.1233404  4.7123890      NA
#> [173,] -3.5342917  4.7123890      NA
#> [174,] -2.9452431  4.7123890      NA
#> [175,] -2.3561945  4.7123890      NA
#> [176,] -1.7671459  4.7123890      NA
#> [177,] -1.1780972  4.7123890      NA
#> [178,] -0.5890486  4.7123890      NA
#> [179,]  0.0000000  4.7123890      NA
#> [180,]  0.5890486  4.7123890      NA
#> [181,]  1.1780972  4.7123890      NA
#> [182,]  1.7671459  4.7123890      NA
#> [183,]  2.3561945  4.7123890      NA
#> [184,]  2.9452431  4.7123890      NA
#> [185,]  3.5342917  4.7123890      NA
#> [186,]  4.1233404  4.7123890      NA
#> [187,]  4.7123890  4.7123890      NA
#>