Smooth an element from a matrix by averaging its neighbors

Smooth an element from a matrix by averaging its neighbors

Usage

smoother(w, period.mat, w.k1, w.k2, b1 = 1, b2 = b1, loo = FALSE, kernel_uni)

Arguments

w

Single frequency to be smoothed.

period.mat

Matrix. The naive spectral estimate (i.e., periodogram), could be complex-valued.

w.k1, w.k2

Vectors containing the whole frequencies at horizontal and vertical directions.

b1, b2

Numeric. Bandwidths at horizontal and vertical direction, respectively.

loo

Logical. If TRUE, conduct leave-one-out kernel smoothing (the center w will be excluded when averaging). Otherwise, keep the center w for averaging.

kernel_uni

Univariate kernel function.