Frequency grid generator

Frequency grid generator

Usage

generate_freq(A1, A2, ext.factor = NULL, return.comb = FALSE, endpt = 1.5)

Arguments

A1, A2

Side lengths of the observational window.

ext.factor

A positive value indicating the extension factor of frequency. If NULL, no extension is conducted.

return.comb

Logical. If TURE, also return the data.frame.

endpt

A positive value indicating the scale factor of the endpoint frequency.

Value

A list of two frequency vectors. If return.comb = TRUE, also return a data.frame containing all combinations of frequencies.

Details

This function generates the frequency grid which is used in periodogram() and periodogram_smooth(). In periodogram_smooth(), to correct the edge effect, the ext.factor argument in generate_freq() is set to 2 by default. Suppose the original frequency ranges from -1.5*π to 1.5*π. Then the extended frequency ranges from -1.5*π*ext.factor to 1.5*π*ext.factor.