Kernel spectral estimator of a multivatiate (multitype) point pattern

Computes the kernel spectral estimator for a multivatiate (multitype) point pattern.

Usage

periodogram_smooth(
  ppp,
  i = NULL,
  j = i,
  inten.formula = "~1",
  data.covariate = NULL,
  bandwidth,
  correct = TRUE,
  a = 0.025,
  A1 = NULL,
  A2 = A1,
  endpt = 1.5,
  equal = TRUE,
  kern = bartlett_uni
)

Arguments

ppp

A point pattern of class "ppp".

i

Mark index. An element in levels(spatstat.geom::marks(ppp)).

j

Mark index. An element in levels(spatstat.geom::marks(ppp)).

inten.formula

A formula syntax in character format specifying the log-liner model for the intensity function, which is passed to ppm(). The default is constant intensity inten.formula = "~1".

data.covariate

Optional. The values of spatial covariates passed to the data argument in ppm.

bandwidth

A positive value indicating the bandwidth of kernel, determined by select_band().

correct

Logical. If TRUE (default), conduct edge correction when computing the kernel spectral estimator.

a

Taper coefficient, a value within unit interval. If a = 1, then taper is not applied. Default is a = 0.025.

A1, A2

Optional. Side lengths of the observation window.

endpt

A positive value indicating the scale factor of the endpoint frequency.

equal

Logical. If TRUE, then use the same bandwidth for both x and y direction.

kern

Univariate scaled kernel function. The default is Barrlett kernel.

Value

A list of matrices, or a single matrix if i and j are specified.

Details

The minimal required arguments are ppp, inten.formula, and bandwidth. If you use any spatial covariate other than the Cartesian coordinates in inten.formula, then data.covariate is also needed. All the other arguments can be left by default setting. periodogram_smooth() computes all the pairwise (marginal and cross-) kernel spectral estimators automatically when the mark indices i and j are unspecified. If i and j are specified, then it only computes the result for that mark combination.

The bandwidth can be determined by select_band().

Examples

library(spatstat)
lam <- function(x, y, m) {(x^2 + y) * ifelse(m == "A", 2, 1)}
spp <- rmpoispp(lambda = lam, win = square(5), types = c("A","B"))

# Compute kernel spectral estimator with intensity fitted by log-linear model
# with Cartesian coordinates
periodogram_smooth(spp, inten.formula = "~ x + y", bandwidth = 1.2)
#> $`A, A`
#>                    -4.71238898038469 -3.76991118430775 -2.82743338823081
#> -4.71238898038469          0.4980021         0.2071694         0.3015873
#> -3.76991118430775          0.3722611         0.2338369         0.7371624
#> -2.82743338823081          0.4228053         0.1936983         0.6220148
#> -1.88495559215388          0.6632194         0.2250690         0.2529196
#> -0.942477796076938         1.1132451         0.5580691         0.4925663
#> 0                          0.7517224         0.3504831         0.5204713
#> 0.942477796076938          0.8327791         0.4660927         0.2284732
#> 1.88495559215388           0.5468418         0.6004535         0.4654435
#> 2.82743338823081           0.5181261         0.7799636         0.6804350
#> 3.76991118430775           0.8228617         0.8107018         0.3579053
#> 4.71238898038469           0.6163748         0.4278463         0.2746250
#>                    -1.88495559215388 -0.942477796076938         0
#> -4.71238898038469          0.5264160          0.7853847 1.2100444
#> -3.76991118430775          0.9280166          0.8623225 1.0499695
#> -2.82743338823081          0.8015703          0.5736285 0.4440405
#> -1.88495559215388          0.4186404          0.6864439 0.5441570
#> -0.942477796076938         0.8893937          1.3435263 0.5641269
#> 0                          0.5454929          0.5044497 0.2610086
#> 0.942477796076938          0.1808627          0.4461855 0.5641269
#> 1.88495559215388           0.1507502          0.2467114 0.5441570
#> 2.82743338823081           0.3125870          0.2869911 0.4440405
#> 3.76991118430775           0.5529940          1.0008897 1.0499695
#> 4.71238898038469           0.7767646          1.0213680 1.2100444
#>                    0.942477796076938 1.88495559215388 2.82743338823081
#> -4.71238898038469          1.0213680        0.7767646        0.2746250
#> -3.76991118430775          1.0008897        0.5529940        0.3579053
#> -2.82743338823081          0.2869911        0.3125870        0.6804350
#> -1.88495559215388          0.2467114        0.1507502        0.4654435
#> -0.942477796076938         0.4461855        0.1808627        0.2284732
#> 0                          0.5044497        0.5454929        0.5204713
#> 0.942477796076938          1.3435263        0.8893937        0.4925663
#> 1.88495559215388           0.6864439        0.4186404        0.2529196
#> 2.82743338823081           0.5736285        0.8015703        0.6220148
#> 3.76991118430775           0.8623225        0.9280166        0.7371624
#> 4.71238898038469           0.7853847        0.5264160        0.3015873
#>                    3.76991118430775 4.71238898038469
#> -4.71238898038469         0.4278463        0.6163748
#> -3.76991118430775         0.8107018        0.8228617
#> -2.82743338823081         0.7799636        0.5181261
#> -1.88495559215388         0.6004535        0.5468418
#> -0.942477796076938        0.4660927        0.8327791
#> 0                         0.3504831        0.7517224
#> 0.942477796076938         0.5580691        1.1132451
#> 1.88495559215388          0.2250690        0.6632194
#> 2.82743338823081          0.1936983        0.4228053
#> 3.76991118430775          0.2338369        0.3722611
#> 4.71238898038469          0.2071694        0.4980021
#> 
#> $`A, B`
#>                            -4.71238898038469        -3.76991118430775
#> -4.71238898038469  -0.002035840+0.102741646i -0.101046914-0.03459951i
#> -3.76991118430775   0.094593870+0.144762090i  0.002452991+0.03252650i
#> -2.82743338823081   0.115249067-0.064443488i  0.077818444+0.07039235i
#> -1.88495559215388   0.114955633-0.143455204i  0.045947789+0.04740057i
#> -0.942477796076938  0.428093578+0.026484105i  0.262643084+0.45739003i
#> 0                   0.174648098-0.009697533i  0.032081125+0.12671648i
#> 0.942477796076938   0.009101848+0.225980324i -0.014349106+0.12731016i
#> 1.88495559215388   -0.006724882+0.280075582i  0.085998643+0.26530226i
#> 2.82743338823081    0.011588967-0.135178418i  0.118816428+0.06174654i
#> 3.76991118430775   -0.203092479-0.224555230i -0.126189354-0.15849954i
#> 4.71238898038469   -0.094932251+0.134096685i -0.250610328+0.04648188i
#>                           -2.82743338823081          -1.88495559215388
#> -4.71238898038469  -0.06961029-0.030683804i -0.2141906830+0.072810867i
#> -3.76991118430775   0.03508154-0.038424074i -0.2259591745-0.071476753i
#> -2.82743338823081   0.20354101+0.091381466i -0.0268236379-0.133565140i
#> -1.88495559215388  -0.01020358+0.008818596i -0.0006761408+0.022413477i
#> -0.942477796076938 -0.13831378+0.285331158i  0.0122845528+0.331005422i
#> 0                  -0.36269908+0.016320599i -0.2679924864+0.155229121i
#> 0.942477796076938  -0.01848792+0.066267719i  0.0239010005+0.088658542i
#> 1.88495559215388    0.17384472+0.091290955i  0.1128822386+0.014810397i
#> 2.82743338823081    0.33659167+0.075854824i  0.1532719961-0.005760237i
#> 3.76991118430775    0.03394692+0.028094606i  0.0191058423+0.064516425i
#> 4.71238898038469   -0.14263209+0.017115663i -0.0532958545-0.040411837i
#>                          -0.942477796076938                       0
#> -4.71238898038469  -0.123111737+0.13324813i -0.04006206-0.06761332i
#> -3.76991118430775  -0.077635606+0.03908725i  0.03078817-0.08641048i
#> -2.82743338823081  -0.191384525-0.11665579i -0.18023418-0.08775172i
#> -1.88495559215388   0.007609142+0.12823697i -0.12288815+0.12346081i
#> -0.942477796076938  0.546588560+0.36949269i  0.12892148+0.20291244i
#> 0                   0.154616324+0.14754432i  0.10214280+0.00000000i
#> 0.942477796076938   0.119737629+0.21422858i  0.12892148-0.20291244i
#> 1.88495559215388   -0.025528938+0.10836718i -0.12288815-0.12346081i
#> 2.82743338823081   -0.048400367+0.07412305i -0.18023418+0.08775172i
#> 3.76991118430775   -0.023646848+0.26720433i  0.03078817+0.08641048i
#> 4.71238898038469   -0.060495737+0.14606005i -0.04006206+0.06761332i
#>                           0.942477796076938           1.88495559215388
#> -4.71238898038469  -0.060495737-0.14606005i -0.0532958545+0.040411837i
#> -3.76991118430775  -0.023646848-0.26720433i  0.0191058423-0.064516425i
#> -2.82743338823081  -0.048400367-0.07412305i  0.1532719961+0.005760237i
#> -1.88495559215388  -0.025528938-0.10836718i  0.1128822386-0.014810397i
#> -0.942477796076938  0.119737629-0.21422858i  0.0239010005-0.088658542i
#> 0                   0.154616324-0.14754432i -0.2679924864-0.155229121i
#> 0.942477796076938   0.546588560-0.36949269i  0.0122845528-0.331005422i
#> 1.88495559215388    0.007609142-0.12823697i -0.0006761408-0.022413477i
#> 2.82743338823081   -0.191384525+0.11665579i -0.0268236379+0.133565140i
#> 3.76991118430775   -0.077635606-0.03908725i -0.2259591745+0.071476753i
#> 4.71238898038469   -0.123111737-0.13324813i -0.2141906830-0.072810867i
#>                            2.82743338823081         3.76991118430775
#> -4.71238898038469  -0.14263209-0.017115663i -0.250610328-0.04648188i
#> -3.76991118430775   0.03394692-0.028094606i -0.126189354+0.15849954i
#> -2.82743338823081   0.33659167-0.075854824i  0.118816428-0.06174654i
#> -1.88495559215388   0.17384472-0.091290955i  0.085998643-0.26530226i
#> -0.942477796076938 -0.01848792-0.066267719i -0.014349106-0.12731016i
#> 0                  -0.36269908-0.016320599i  0.032081125-0.12671648i
#> 0.942477796076938  -0.13831378-0.285331158i  0.262643084-0.45739003i
#> 1.88495559215388   -0.01020358-0.008818596i  0.045947789-0.04740057i
#> 2.82743338823081    0.20354101-0.091381466i  0.077818444-0.07039235i
#> 3.76991118430775    0.03508154+0.038424074i  0.002452991-0.03252650i
#> 4.71238898038469   -0.06961029+0.030683804i -0.101046914+0.03459951i
#>                             4.71238898038469
#> -4.71238898038469  -0.094932251-0.134096685i
#> -3.76991118430775  -0.203092479+0.224555230i
#> -2.82743338823081   0.011588967+0.135178418i
#> -1.88495559215388  -0.006724882-0.280075582i
#> -0.942477796076938  0.009101848-0.225980324i
#> 0                   0.174648098+0.009697533i
#> 0.942477796076938   0.428093578-0.026484105i
#> 1.88495559215388    0.114955633+0.143455204i
#> 2.82743338823081    0.115249067+0.064443488i
#> 3.76991118430775    0.094593870-0.144762090i
#> 4.71238898038469   -0.002035840-0.102741646i
#> 
#> $`B, B`
#>                    -4.71238898038469 -3.76991118430775 -2.82743338823081
#> -4.71238898038469          0.1388236         0.2477936        0.09557042
#> -3.76991118430775          0.2585956         0.1166373        0.08751592
#> -2.82743338823081          0.4047750         0.2226474        0.23529730
#> -1.88495559215388          0.2322113         0.3768464        0.28028390
#> -0.942477796076938         0.4171760         0.9604458        0.49694778
#> 0                          0.2026087         0.4566886        0.53386081
#> 0.942477796076938          0.1848609         0.1637697        0.40925244
#> 1.88495559215388           0.3152459         0.2109711        0.22041899
#> 2.82743338823081           0.3194821         0.1598477        0.24678180
#> 3.76991118430775           0.2641902         0.1750373        0.16951055
#> 4.71238898038469           0.2407151         0.2868946        0.21066647
#>                    -1.88495559215388 -0.942477796076938          0
#> -4.71238898038469          0.1810707         0.12757357 0.07248645
#> -3.76991118430775          0.1427475         0.08921327 0.08662322
#> -2.82743338823081          0.1262671         0.20333496 0.24972002
#> -1.88495559215388          0.1099406         0.17349239 0.21629836
#> -0.942477796076938         0.2991442         0.44698011 0.33943552
#> 0                          0.4562856         0.24510838 0.16064543
#> 0.942477796076938          0.4889129         0.35434211 0.33943552
#> 1.88495559215388           0.4197307         0.29381400 0.21629836
#> 2.82743338823081           0.1942024         0.14780681 0.24972002
#> 3.76991118430775           0.0717761         0.11526788 0.08662322
#> 4.71238898038469           0.0629095         0.06600637 0.07248645
#>                    0.942477796076938 1.88495559215388 2.82743338823081
#> -4.71238898038469         0.06600637        0.0629095       0.21066647
#> -3.76991118430775         0.11526788        0.0717761       0.16951055
#> -2.82743338823081         0.14780681        0.1942024       0.24678180
#> -1.88495559215388         0.29381400        0.4197307       0.22041899
#> -0.942477796076938        0.35434211        0.4889129       0.40925244
#> 0                         0.24510838        0.4562856       0.53386081
#> 0.942477796076938         0.44698011        0.2991442       0.49694778
#> 1.88495559215388          0.17349239        0.1099406       0.28028390
#> 2.82743338823081          0.20333496        0.1262671       0.23529730
#> 3.76991118430775          0.08921327        0.1427475       0.08751592
#> 4.71238898038469          0.12757357        0.1810707       0.09557042
#>                    3.76991118430775 4.71238898038469
#> -4.71238898038469         0.2868946        0.2407151
#> -3.76991118430775         0.1750373        0.2641902
#> -2.82743338823081         0.1598477        0.3194821
#> -1.88495559215388         0.2109711        0.3152459
#> -0.942477796076938        0.1637697        0.1848609
#> 0                         0.4566886        0.2026087
#> 0.942477796076938         0.9604458        0.4171760
#> 1.88495559215388          0.3768464        0.2322113
#> 2.82743338823081          0.2226474        0.4047750
#> 3.76991118430775          0.1166373        0.2585956
#> 4.71238898038469          0.2477936        0.1388236
#> 
#> attr(,"inten.fitted.scaled")
#> attr(,"inten.fitted.scaled")$A
#> function(x,y) of class ‘funxy’
#> window: binary image mask
#> 128 x 128 pixel array (ny, nx)
#> enclosing rectangle: [-0.5, 0.5] x [-0.5, 0.5] units
#> 
#> Original function definition:
#> function (x, y) 
#> {
#>     Z[list(x = x, y = y)]
#> }
#> <bytecode: 0x55d45652cf20>
#> <environment: 0x55d4567cd828>
#> 
#> attr(,"inten.fitted.scaled")$B
#> function(x,y) of class ‘funxy’
#> window: binary image mask
#> 128 x 128 pixel array (ny, nx)
#> enclosing rectangle: [-0.5, 0.5] x [-0.5, 0.5] units
#> 
#> Original function definition:
#> function (x, y) 
#> {
#>     Z[list(x = x, y = y)]
#> }
#> <bytecode: 0x55d45652cf20>
#> <environment: 0x55d4565ad420>
#> 
#> attr(,"a")
#> [1] 0.025
#> attr(,"A1")
#> [1] 5
#> attr(,"A2")
#> [1] 5
#> attr(,"freq.list")
#> attr(,"freq.list")$omega1
#>  [1] -4.7123890 -3.7699112 -2.8274334 -1.8849556 -0.9424778  0.0000000
#>  [7]  0.9424778  1.8849556  2.8274334  3.7699112  4.7123890
#> 
#> attr(,"freq.list")$omega2
#>  [1] -4.7123890 -3.7699112 -2.8274334 -1.8849556 -0.9424778  0.0000000
#>  [7]  0.9424778  1.8849556  2.8274334  3.7699112  4.7123890
#> 
#> attr(,"freq.list")$omega.comb
#>            omega1     omega2 Density
#>   [1,] -4.7123890 -4.7123890      NA
#>   [2,] -3.7699112 -4.7123890      NA
#>   [3,] -2.8274334 -4.7123890      NA
#>   [4,] -1.8849556 -4.7123890      NA
#>   [5,] -0.9424778 -4.7123890      NA
#>   [6,]  0.0000000 -4.7123890      NA
#>   [7,]  0.9424778 -4.7123890      NA
#>   [8,]  1.8849556 -4.7123890      NA
#>   [9,]  2.8274334 -4.7123890      NA
#>  [10,]  3.7699112 -4.7123890      NA
#>  [11,]  4.7123890 -4.7123890      NA
#>  [12,] -4.7123890 -3.7699112      NA
#>  [13,] -3.7699112 -3.7699112      NA
#>  [14,] -2.8274334 -3.7699112      NA
#>  [15,] -1.8849556 -3.7699112      NA
#>  [16,] -0.9424778 -3.7699112      NA
#>  [17,]  0.0000000 -3.7699112      NA
#>  [18,]  0.9424778 -3.7699112      NA
#>  [19,]  1.8849556 -3.7699112      NA
#>  [20,]  2.8274334 -3.7699112      NA
#>  [21,]  3.7699112 -3.7699112      NA
#>  [22,]  4.7123890 -3.7699112      NA
#>  [23,] -4.7123890 -2.8274334      NA
#>  [24,] -3.7699112 -2.8274334      NA
#>  [25,] -2.8274334 -2.8274334      NA
#>  [26,] -1.8849556 -2.8274334      NA
#>  [27,] -0.9424778 -2.8274334      NA
#>  [28,]  0.0000000 -2.8274334      NA
#>  [29,]  0.9424778 -2.8274334      NA
#>  [30,]  1.8849556 -2.8274334      NA
#>  [31,]  2.8274334 -2.8274334      NA
#>  [32,]  3.7699112 -2.8274334      NA
#>  [33,]  4.7123890 -2.8274334      NA
#>  [34,] -4.7123890 -1.8849556      NA
#>  [35,] -3.7699112 -1.8849556      NA
#>  [36,] -2.8274334 -1.8849556      NA
#>  [37,] -1.8849556 -1.8849556      NA
#>  [38,] -0.9424778 -1.8849556      NA
#>  [39,]  0.0000000 -1.8849556      NA
#>  [40,]  0.9424778 -1.8849556      NA
#>  [41,]  1.8849556 -1.8849556      NA
#>  [42,]  2.8274334 -1.8849556      NA
#>  [43,]  3.7699112 -1.8849556      NA
#>  [44,]  4.7123890 -1.8849556      NA
#>  [45,] -4.7123890 -0.9424778      NA
#>  [46,] -3.7699112 -0.9424778      NA
#>  [47,] -2.8274334 -0.9424778      NA
#>  [48,] -1.8849556 -0.9424778      NA
#>  [49,] -0.9424778 -0.9424778      NA
#>  [50,]  0.0000000 -0.9424778      NA
#>  [51,]  0.9424778 -0.9424778      NA
#>  [52,]  1.8849556 -0.9424778      NA
#>  [53,]  2.8274334 -0.9424778      NA
#>  [54,]  3.7699112 -0.9424778      NA
#>  [55,]  4.7123890 -0.9424778      NA
#>  [56,] -4.7123890  0.0000000      NA
#>  [57,] -3.7699112  0.0000000      NA
#>  [58,] -2.8274334  0.0000000      NA
#>  [59,] -1.8849556  0.0000000      NA
#>  [60,] -0.9424778  0.0000000      NA
#>  [61,]  0.0000000  0.0000000      NA
#>  [62,]  0.9424778  0.0000000      NA
#>  [63,]  1.8849556  0.0000000      NA
#>  [64,]  2.8274334  0.0000000      NA
#>  [65,]  3.7699112  0.0000000      NA
#>  [66,]  4.7123890  0.0000000      NA
#>  [67,] -4.7123890  0.9424778      NA
#>  [68,] -3.7699112  0.9424778      NA
#>  [69,] -2.8274334  0.9424778      NA
#>  [70,] -1.8849556  0.9424778      NA
#>  [71,] -0.9424778  0.9424778      NA
#>  [72,]  0.0000000  0.9424778      NA
#>  [73,]  0.9424778  0.9424778      NA
#>  [74,]  1.8849556  0.9424778      NA
#>  [75,]  2.8274334  0.9424778      NA
#>  [76,]  3.7699112  0.9424778      NA
#>  [77,]  4.7123890  0.9424778      NA
#>  [78,] -4.7123890  1.8849556      NA
#>  [79,] -3.7699112  1.8849556      NA
#>  [80,] -2.8274334  1.8849556      NA
#>  [81,] -1.8849556  1.8849556      NA
#>  [82,] -0.9424778  1.8849556      NA
#>  [83,]  0.0000000  1.8849556      NA
#>  [84,]  0.9424778  1.8849556      NA
#>  [85,]  1.8849556  1.8849556      NA
#>  [86,]  2.8274334  1.8849556      NA
#>  [87,]  3.7699112  1.8849556      NA
#>  [88,]  4.7123890  1.8849556      NA
#>  [89,] -4.7123890  2.8274334      NA
#>  [90,] -3.7699112  2.8274334      NA
#>  [91,] -2.8274334  2.8274334      NA
#>  [92,] -1.8849556  2.8274334      NA
#>  [93,] -0.9424778  2.8274334      NA
#>  [94,]  0.0000000  2.8274334      NA
#>  [95,]  0.9424778  2.8274334      NA
#>  [96,]  1.8849556  2.8274334      NA
#>  [97,]  2.8274334  2.8274334      NA
#>  [98,]  3.7699112  2.8274334      NA
#>  [99,]  4.7123890  2.8274334      NA
#> [100,] -4.7123890  3.7699112      NA
#> [101,] -3.7699112  3.7699112      NA
#> [102,] -2.8274334  3.7699112      NA
#> [103,] -1.8849556  3.7699112      NA
#> [104,] -0.9424778  3.7699112      NA
#> [105,]  0.0000000  3.7699112      NA
#> [106,]  0.9424778  3.7699112      NA
#> [107,]  1.8849556  3.7699112      NA
#> [108,]  2.8274334  3.7699112      NA
#> [109,]  3.7699112  3.7699112      NA
#> [110,]  4.7123890  3.7699112      NA
#> [111,] -4.7123890  4.7123890      NA
#> [112,] -3.7699112  4.7123890      NA
#> [113,] -2.8274334  4.7123890      NA
#> [114,] -1.8849556  4.7123890      NA
#> [115,] -0.9424778  4.7123890      NA
#> [116,]  0.0000000  4.7123890      NA
#> [117,]  0.9424778  4.7123890      NA
#> [118,]  1.8849556  4.7123890      NA
#> [119,]  2.8274334  4.7123890      NA
#> [120,]  3.7699112  4.7123890      NA
#> [121,]  4.7123890  4.7123890      NA
#> 
#> attr(,"cate")
#> [1] "A" "B"
#> attr(,"cate.comb")
#>      pos1 pos2
#> [1,] "A"  "A" 
#> [2,] "A"  "B" 
#> [3,] "B"  "B" 
#> attr(,"bandwidth")
#>  b1  b2 
#> 1.2 1.2